+++ eslint.config.mjs
@@ -4,13 +4,11 @@ import tseslint from "typescript-eslint";
 import { defineConfig, globalIgnores } from "eslint/config";
 
 export default defineConfig([
-  globalIgnores(["dist"]),
   {
-    files: ["**/*.ts"],
+    files: ["**/*.{js,ts}"],
     extends: [js.configs.recommended, tseslint.configs.recommended],
-    languageOptions: {
-      ecmaVersion: 2020,
-      globals: globals.browser,
-    },
+    languageOptions: { globals: globals.browser },
   },
+  tseslint.configs.recommended,
+  globalIgnores(["node_modules/**", "dist/**", "public/**"]),
 ]);