+++ changelog.md
@@ -16,6 +16,8 @@
 - Icons:
   - Brand icons migrated to Simple Icons (No longer the same style either)
   - Remove Linkedin link
+- Gameplay:
+  - Vertically center game canvas
 
 # 0.8.12 Beta - 20/03/2026 23:30 UTC+1