+++ changelog.md
@@ -11,6 +11,7 @@
     `dialog-target` and `dialog-action` attributes replacing `closeButtonIds` array and custom logic
   - `LastGameStore`: Remove `hasData()` so checks now rely directly on the `store` property
   - Simplify filenames for all bundles and update `precacheResources.json`
+  - Simplify `mergeHtml` Vite plugin and add build date to generated HTML
 - Authentication:
   - Implement Facebook SSO
 - Settings: